pyproject.toml
[project]
name = "gradio-pyinstaller-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.3.0",
    "pywebview>=5.3.2",
]

[tool.uv]
dev-dependencies = [
    "pyinstaller>=6.11.1",
]
